UserInfo API

+++ src/com/simplebbs/po/Report.java
@@ -3,18 +3,18 @@ package com.simplebbs.po;
 import java.util.Date;
 
 public class Report {
-    private long record_id;
+    private int record_id;
     private long post_id;
     private long comment_id;
     private int reporter;
     private String report_reason;
     private Date report_time;
 
-    public void setRecord_id(long record_id) {
+    public void setRecord_id(int record_id) {
         this.record_id = record_id;
     }
 
-    public long getRecord_id() {
+    public int getRecord_id() {
         return record_id;
     }
